Correct touchscreen coordinates
W3School tutorial incorrectly used touch event screenX instead of pageX
and didn't compute the offset between canvas and its parent (body). This caused
the gamePiece to not display at all on touch-based devices, until the touch was
complete and the mousemove handler kicked in.

Code snippet to correct the offset:
https://stackoverflow.com/questions/17165382/getting-the-correct-coordinates-of-a-canvas-touch-event